move all the things to where they belong
