+++ stubs/cachetools/cachetools/fifo.pyi
@@ -1,0 +1,2 @@
+# this module is deprecated
+from . import FIFOCache as FIFOCache
