+++ stubs/cachetools/cachetools/mru.pyi
@@ -1,0 +1,2 @@
+# this module is deprecated
+from . import TTLCache as TTLCache
